+++ RELEASES.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 vNext
 -----
 
+v0.7.0 (2018-06-17)
+-------------------
+
 ### aomx
 
 - Improved: GoDoc (added code example, improved existing comments)